Move kernel and topo tests to their proper unittest subdirectories
