+++ .github/workflows/unit-test.yml
@@ -83,18 +83,18 @@ jobs:
           cd /__w/FlagCX/FlagCX/test/unittest/runner
           export MPI_HOME=/usr/local/mpi
           export PATH=$MPI_HOME/bin:$PATH
-          mpirun -np 8 --allow-run-as-root ./build/bin/runner_mpi_tests --gtest_filter=-FlagCXKernelTest.*
+          mpirun -np 8 --allow-run-as-root ./build/bin/runner_mpi_tests
 
       - name: Run multi-cluster tests with mpirun
         run: |
           cd /__w/FlagCX/FlagCX/test/unittest/runner
           export MPI_HOME=/usr/local/mpi
           export PATH=$MPI_HOME/bin:$PATH
-          mpirun -np 8 --allow-run-as-root -x FLAGCX_MEM_ENABLE=1 -x FLAGCX_CLUSTER_SPLIT_LIST=2 -x FLAGCX_IB_HCA=mlx5 ./build/bin/runner_mpi_tests --gtest_filter=-FlagCXKernelTest.*
+          mpirun -np 8 --allow-run-as-root -x FLAGCX_MEM_ENABLE=1 -x FLAGCX_CLUSTER_SPLIT_LIST=2 -x FLAGCX_IB_HCA=mlx5 ./build/bin/runner_mpi_tests
 
       - name: Run kernel tests with mpirun
         run: |
-          cd /__w/FlagCX/FlagCX/test/unittest/runner
+          cd /__w/FlagCX/FlagCX/test/unittest/kernel
           export MPI_HOME=/usr/local/mpi
           export PATH=$MPI_HOME/bin:$PATH
-          mpirun -np 8 --allow-run-as-root -x FLAGCX_MEM_ENABLE=1 -x FLAGCX_CLUSTER_SPLIT_LIST=2 -x FLAGCX_USE_HETERO_COMM=1 ./build/bin/runner_mpi_tests --gtest_filter=FlagCXKernelTest.*
+          mpirun -np 8 --allow-run-as-root -x FLAGCX_MEM_ENABLE=1 -x FLAGCX_CLUSTER_SPLIT_LIST=2 -x FLAGCX_USE_HETERO_COMM=1 ./build/bin/kernel_mpi_tests
