--- test/unittest/runner/coll_topo_detection.cpp
@@ -1,12 +1,0 @@
-// Topology detection test (migrated from test/unittest/main.cpp)
-#include "runner_fixtures.hpp"
-#include <iostream>
-
-TEST_F(FlagCXTopoTest, TopoDetection) {
-  flagcxComm_t &comm = handler->comm;
-  flagcxUniqueId_t &uniqueId = handler->uniqueId;
-
-  std::cout << "executing flagcxCommInitRank" << std::endl;
-  auto result = flagcxCommInitRank(&comm, nranks, uniqueId, rank);
-  EXPECT_EQ(result, flagcxSuccess);
-}
