+++ test/unittest/runner/include/runner_fixtures.hpp
@@ -32,33 +32,3 @@ protected:
   size_t size;
   size_t count;
 };
-
-class FlagCXTopoTest : public FlagCXTest {
-protected:
-  void SetUp() override;
-  void TearDown() override;
-
-  flagcxHandlerGroup_t handler;
-  flagcxStream_t stream;
-  void *sendbuff;
-  void *recvbuff;
-  void *hostsendbuff;
-  void *hostrecvbuff;
-  size_t size;
-  size_t count;
-};
-
-class FlagCXKernelTest : public FlagCXTest {
-protected:
-  void SetUp() override;
-  void TearDown() override;
-
-  flagcxHandlerGroup_t handler;
-  flagcxStream_t stream;
-  void *sendbuff;
-  void *recvbuff;
-  void *hostsendbuff;
-  void *hostrecvbuff;
-  size_t size;
-  size_t count;
-};
